fix: Fix list sort with groupby (#3990)
## Summary

```
(
    daft.from_pydict(
        {"group_col": [1, 1, 1, 2, 2], "id_col": ["c", "a", "b", "e", "d"]}
    )
    .groupby("group_col")
    .agg(daft.col("id_col").agg_list().list.sort().alias("ids_col"))
    .show()
)
```
was reported as failing with `DaftCoreException: DaftError::ValueError
Need at least 1 series to perform concat`

This happens when the number of partitions in our groupby is greater
than the actual number of partitions needed. This results in some empty
partitions with nothing to sort. We should handle these cases
accordingly.

The same happens for fixed size lists.

This PR fixes the issue.

## Related Issues

Closes #3969 

## Checklist

- [x] All tests have passed