feat: offset indices in sparse tensor (#3725)
Change the indices of sparse tensors from positions of non-zero elements
to offsets.

For example, if we have a tensor:

[0, 29, 0, 0, 17, 6]

its sparse representation would be:

- **values:** 29,17,6
- **indices:** 1,4,5

Instead, we want to store the differences (offsets) between consecutive
indices:

- **values:** 29,17,6
- **indices:** 1,3,1

This approach seems to significantly reduce the size of compressed
Parquet files, as the indices diffs contain repetitive patterns, which
compress well.

---------

Co-authored-by: Itzhak Stern <itzhaks@ubup039.me-corp.lan>
Co-authored-by: Colin Ho <colin.ho99@gmail.com>