fix: surface non 'NotFound' errors during table resolution (#4073)
## Changes Made

This PR introduces a daft "NotFoundError" which is descriptive for the
customers, but also let's the session (via read_table) ignore this while
it looks up the chain of possible table locations. This is necessary
because we support unqualified, schema-qualified, and catalog-qualified
identifiers **with** nested namespaces, so we check that chain in that
order and return NotFound if all three fail.

The previous logical treated all errors as NotFound which was swallowing
connection errors.

## Related Issues

#3984 

## Checklist

- [n/a] Documented in API Docs (if applicable)
- [n/a] Documented in User Guide (if applicable)
- [n/a] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [n/a] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)