chore(optimizer): move join type simplification into new rule (#4122)
## Changes Made

Quick refactor of some of the logic that previously lived in
PushDownFilter into its own rule.

I previously implemented it in PushDownFilter because it could have
interactions with the other filter pushdown logic. However, we should
try to keep PushDownFilter as simple as we can, just for pushing filter
nodes through a plan. Thus, I am moving this logic into a new rule and
using the repeated execution of the rule batch to deal with interactions
between the two rules.

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

Related to why we closed #4118

## Checklist

- [x] Documented in API Docs (if applicable)
- [x] Documented in User Guide (if applicable)
- [x] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [x] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)