fix(join ordering): Fix bug with aliasing columns between joins (#4133)
## Changes Made

In #4113 we have an alias in between join nodes. The join graph
constructor has a bug where aliases of non-unique columns across
different query tree branches would all be aliased. This is incorrect
behaviour.

As a quick fix, we avoid reordering projects in between join nodes if
the projection has any non-column-reference. A longer term fix (tracked
in #4131) is to add projection and filter pushups during join ordering.

## Related Issues

Closes #4113