feat: add `unix_timestamp` function (#4130)
## Changes Made

adds `unix_timestamp` function to get the number of <timeunits> since
unix epoch.

## Note for reviewers
the OP called this `timestamp` but I thought that was a bit misleading,
and `unix_timestamp` was a bit more explicit.

---

for the implementation, it uses our current casting logic to convert the
date or datetime in to a numeric value


Usage:

- `col('date').dt.unix_timestamp()` -> milliseconds since epoch
or pass in a time unit str
- `col('date').dt.unix_timestamp('s')` -> seconds since epoch
or pass in a time unit object
- `col('date').dt.unix_timestamp(TimeUnit.s())` -> seconds since epoch



## Related Issues

https://github.com/Eventual-Inc/Daft/issues/4094

## Checklist

- [x] Documented in API Docs (if applicable)
- [n/a] Documented in User Guide (if applicable)
- [n/a] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [n/a] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)