feat: adds case-sensitive and case-normalized catalog and table identifiers [1/3] (#4149)
## Changes Made

This PR extends the bindings to support both case-sensitive and
case-normalized identifier handling for attached catalogs and tables. It
does not modify column resolution (yet) and it does not add
case-insensitive matching (yet). This PR also includes documentation for
the goal state which will be [linked once
built](https://www.getdaft.io/projects/docs/en/rchowell-identifiers/sql/identifiers/).
I have also implemented case-insensitive for the attached catalogs and
attached tables, but it is not possible to use yet since I have the
exact-case hardcoded as default as to not break things.


[Documentation](https://www.getdaft.io/projects/docs/en/rchowell-identifiers/sql/identifiers/)

**Whats next?**

* Case-insensitive as default mode.
* Session options with SET statement.
* Session option for identifier mode.
* Column resolution with the modes.

## Related Issues

* Closes #3765 

## Checklist

- [ ] Documented in API Docs (if applicable)
- [x] Documented in User Guide (if applicable)
- [x] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [x] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)