feat: add `strftime` (#4146)
## Changes Made

- adds `to_string`/`strftime`. 

### Note for reviewers. 

on the python side I added both `to_string` and `strftime`, but on sql I
only added `strftime`.

The reasoning is that because we don't yet have namespaced prefixes on
sql, so `to_string` is pretty generic, but on python it's `dt.to_string`
which fits with our `str.to_date` and `str.to_datetime`.



## Related Issues

closes https://github.com/Eventual-Inc/Daft/issues/4094

## Checklist

- [x] Documented in API Docs (if applicable)
- [n/a] Documented in User Guide (if applicable)
- [n/a] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [n/a] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)