refactor(ordinals): relax unique column name constraint in schema (#4204)
## Changes Made

As a first step in our ordinals refactor, we want to relax the
constraint on the Daft schema that all column names must be unique. We
do this by modifying our schema struct to now hold a vec of fields
instead of an indexmap (plus a multimap of names to indices for
convenience).

One effect of this is that our schema constructor no longer errors. I
also made these fields private and added helper functions for using the
schema instead.

The above is done in `src/daft-schema/src/schema.rs`. Everything else
should naturally from there, all behavior is preserved.

## Related Issues

none

## Checklist

- [x] Documented in API Docs (if applicable)
- [x] Documented in User Guide (if applicable)
- [x] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [x] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)