fix: simplify and fix creating tables from dataframes (#4196)
## Changes Made

This PR simplifies how create_table works. Originally we had a source
object with the intention of support diverse arguments as the table
source, but in practice it's either a schema or dataframe. Also, we
still need to model this union on the rust side so PyTableSource still
exists as the argument for create_temp_table. This may warrant a revisit
once Session logic makes its way into rust, but for now this
simplification is better documented and fixes. The customers need not be
aware of "TableSource" since we will always be able to construct these
behind the scenes. Similarly, if we wish to support diverse table
sources, we are likely better of extending DataFrame creation as to not
complicate this.

## Related Issues

- Closes #4177 

## Checklist

- [x] Documented in API Docs (if applicable)
- [n/a] Documented in User Guide (if applicable)
- [n/a] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [n/a] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)