feat: defines pushdown terms for python (#4203)
## Changes Made

This introduces a simple Term IR ([much like
pyiceberg](https://py.iceberg.apache.org/reference/pyiceberg/expressions/#pyiceberg.expressions.Term))
which is a model for pushdown expressions. This is scoped to the python
pushdown package, and the Term IR comes with a visitor which can be used
to translate to other domains. This enable easy consumption of PyExpr /
rust expressions. I have an example visitor which produces
s-expressions, and can add one to translate to the PyIceberg Expression
IR.

An alternative design is no IR and visitor-only - this is nice because
there is no additional IR, however it comes at the tradeoff of possible
tighter coupling to the Rust expressions and we may want an IR anyways
if we ever want to serde pushdown expressions. Please let me know what
you all think is best here. #4210