fix: Allow PNGs to be displayed notebooks (#4255)
## Changes Made

When we generate thumbnails to display in notebooks, we encoded them as
JPEG by default. This does not work if images have an alpha channel.

This PR changes it so that we generate PNGs instead. This should have
minimal performance impact because we're generating small 128x128
thumbnails and this only happens when previewing a small amount
(typically <= 8) of results in a notebook.

## Related Issues

Closes #4072