chore: Upgrade supported pyarrow version to 19.0.0 (#4257)
## Changes Made

Upgrades our pyarrow version in tests to 19.0.0, additionally removes
the pin in `pyproject.toml`

In order to support 19.0.0, this PR also fixes our map type conversion
code, which is the only thing that failed. See
https://github.com/Eventual-Inc/Daft/issues/4206

Map types are required to be annotated with `not null` for the keys, and
currently this is not propagated from daft datatype to arrow. This PR
modifies the `to_arrow` conversion for map types and arrays to obey this
constraint. Previous version of pyarrow did not have this error as they
auto-added the annotation when importing from the C data interface.


## Related Issues

Closes https://github.com/Eventual-Inc/Daft/issues/4256

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)