refactor(ordinals): add bound column variant (#4269)
## Changes Made

Created a `BoundColumn` column variant that holds an ordinal/index and a
field.

This is not currently used anywhere except in RecordBatch for evaluating
an expression. The plan is to move the binding step up the
planning/execution pipeline, until it completely replaces the Resolved
variant. That way, we can do this refactor incrementally and ensure that
tests pass the whole time. See [the
roadmap](https://github.com/Eventual-Inc/Daft/issues/4270) for more
details

## Related Issues

#4270

## Checklist

- [x] Documented in API Docs (if applicable)
- [x] Documented in User Guide (if applicable)
- [x] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [x] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)