ci: Bump pyarrow version in CI to 19.0.1 (#4301)
## Changes Made

Arrow 19.0.0 is unable to read parquet files written by newer versions
of arrow with statistics.

The issue arises from an optimization where level histograms in
statistics (e.g. the level histogram used for repetition levels) is
omitted when the max level is 0.

For more details, see https://github.com/apache/arrow/issues/45283

At the same time we can't bump to pyarrow 20.0.0 because this is
incompatible with ray, see
https://github.com/ray-project/ray/issues/52685

This means we should use the minor release of arrow 19, i.e. arrow
19.0.1