refactor(ordinals): use indices in RecordBatch::get_column(s) (#4318)
## Changes Made

Follow-up to #4297. This should complete the ordinals refactor for
RecordBatch, both on Rust and Python, by removing the ability to
directly get columns by name, and replacing calls to deprecated schema
methods.

The only place in daft-recordbatch that still uses name references are
in the Python boundary and the join kernels, because it does common
column merging. We want to change that behavior in the future anyway, so
I'm leaving it in for now.

Also renamed `MicroPartition.get_column` to
`MicroPartition.get_column_by_name` and added
`MicroPartition.get_column` and `MicroPartition.columns` for consistency
between RecordBatch and MicroPartition APIs. There are still a number of
places using `MicroPartition.get_column_by_name` that I plan to address
when the refactor reaches micropartition

## Related Issues

#4270

## Checklist

- [x] Documented in API Docs (if applicable)
- [x] Documented in User Guide (if applicable)
- [x] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [x] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)