feat: adds expression visitor (#4278)
## Changes Made

This PR adds an `Expression` visitor for traversing, introspecting, and
translating Expression trees. This can be used to translate daft's
pushdowns to a python-based scan task. This works by writing the visitor
interface in python, then implementing the accept in rust. I've
implement it such that the Expression is not a traversable tree, but I
am able to reconstruct the original arguments so implementors can
control a walk if they wish.

* Adds ExpressionVisitor with parameterized return.
* Adds PyExpr `_eq` , `_ne` and `_hash` using the underlying derived
impls
* Please see the TracingVisitor has an example of how to translate aka
traverse + tree fold

## Related Issues

This functionality unlocks and addresses several issues.

- #4239
- https://github.com/Eventual-Inc/Daft/pull/4203
- https://github.com/Eventual-Inc/Daft/issues/1953
- https://github.com/Eventual-Inc/Daft/issues/1800
- https://github.com/Eventual-Inc/Daft/blob/main/daft/io/_lance.py#L115
-
https://github.com/Eventual-Inc/Daft/blob/main/daft/iceberg/iceberg_scan.py#L210-L214
-
https://github.com/Eventual-Inc/Daft/blob/main/daft/hudi/hudi_scan.py#L138-L139

## Checklist

- [x] Documented in API Docs (if applicable)
- [x] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)