docs: add algolia search, separate primary & secondary nav, stylistic changes (#4330)
## Changes Made

### 1. Add Algolia for search

https://github.com/Eventual-Inc/Daft/issues/2496

### 2. Separate primary & secondary nav

Previously, we had one nav on the left but as we add more doc pages,
it's not sustainable because the scroll becomes really long. I've
disabled `toc.integrate` to separate the primary and secondary nav with
some stylistic changes to remove the focus from the secondary nav and
lessen the noise.

This is especially nice for:
1. Beginners trying to navigate through `Quickstart` and `Core Concepts`
pages without crowding the primary nav
2. API docs because the scroll becomes difficult to navigate when
methods are listed

### 3. Smaller Changes

- Added js such that all _external_ links will open in a new tab
- Added Slack button in header
- Added link to Release Notes in nav
- Moved "Ask AI" widget to left side to avoid overlap with secondary nav
- Centered footer content to avoid overlap with "Ask AI" widget and
readthedocs flyout

See build here:
https://www.getdaft.io/projects/docs/en/chanchan-docs-menu/