refactor(ordinals): bound expressions in table statistics (#4342)
## Changes Made

Moved the column binding boundary above the daft-stats crate. In
addition, deprecate `cast_to_schema` and `cast_to_schema_with_fill` for
MicroPartition, RecordBatch, and TableStatistics so that we do not use
them in future code.

Not touching PartitionSpec because it's only used for reads, which I do
not plan on refactoring anyway.

This is a prerequisite for moving the refactor above pushdowns and
micropartition.

## Related Issues

#4270

## Checklist

- [x] Documented in API Docs (if applicable)
- [x] Documented in User Guide (if applicable)
- [x] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [x] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)