fix: CSV read with disjoint predicate pushdown (#4363)
## Changes Made

Previous ordinals refactor PR #4297 caused a regression since the wrong
schema was used to bind predicate pushdown columns in the CSV read. This
causes failures when a column in the predicate is not in the final
projection schema.

This wasn't caught due to a mistaken change in #3804 that disabled the
relevant tests.

Also changed the corresponding JSON read logic because I'm unsure if the
previous code was correct either.

## Related Issues

none

## Checklist

- [x] Documented in API Docs (if applicable)
- [x] Documented in User Guide (if applicable)
- [x] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [x] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)