fix: ignores tests/integration for make test target (#4365)
## Changes Made

Integ tests using torch were added, but the dep was not included in
reqs. This meant I could not `make test` a fresh repo since you'd get a
missing dep error. This has not failed in CI because we scope pytest to
`tests/` in CI, but not in `make test`. Possible solutions are lazy-dep,
scope `make test` down, or add this dep.

## Related Issues

n/a


## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)