chore(exprs): add helper methods for extracting scalars from function args (#4351)
## Changes Made

Adds an `extract` and `extract_optional` to get scalar values from
`FunctionArgs`.

Since we pass some complex types through `lit` such as `IOConfig`, it is
quite a burden to define a mapping between every type. Since we use
serde extensively, it makes a lot of sense to leverage that to handle
this for us. so now we can call the new `make_literal` function on
anything that implements `Serialize`, and retrieve it back out via
`T::try_from_literal`




Under the hood it uses serde to serialize and deserialize any value into
a `LiteralValue`.

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)