feat: Flotilla scheduler (#4349)
## Changes Made

This PR implements the `scheduler` trait, as well as stubbed
implementations for the `Default` scheduler and `Linear` Scheduler.

This PR also refactors several components into generics, namely task and
worker. This will allow us to create different implementations of tasks
and workers, for example mock tasks and mock workers for testing.

I will implement the logic as well as unit tests for scheduler in a
separate PR.

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)