perf: Morsel size ranges for project and filter operators (#4344)
## Changes Made

Rather than requiring that project and filter operators process a morsel
of an exact size, this PR makes the morsel sizing a range and changes it
for project and filters to be any morsel of size 0 to the default. Thus,
these ops will split morsels that are too big but wont group morsels
that are too small anymore. This change affects other operators but
their defaults are unchanged (by making their range 1 value).

Tested with TPCH and ClickBench with +-5% affect on all queries, and a
customer map-style pipeline to no change.

This is 1 of 3/4 PRs to improve the streaming latency of workloads.

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)