refactor(exprs): literal support for FunctionArgs proc macro (#4401)
## Changes Made

Add literal support for the FunctionArgs proc macro. 

Now, instead of using serde to serialize a Rust struct into a
LiteralValue for Python functions, we can just store the Python objects
directly as a `LiteralValue::Python`, and then implement the extraction
to Rust struct in `FromLiteral`. This allows us to remove the
LiteralValue serialization functionality, and also eliminates the need
to special case every type in the Python-exposed `lit` function.

## Related Issues

follow-up on #4348

## Checklist

- [x] Documented in API Docs (if applicable)
- [x] Documented in User Guide (if applicable)
- [x] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [x] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)