feat: enables scalar function lowering e.g. function overloads (#4431)
## Changes Made


* FunctionModule is defined over the ScalarFunctionFactory.
* ScalarFunctionFactory provides a concrete implementation based upon
the given `FunctionArgs<Field>`
* This enables planning time resolution to a concrete implementation.
* I've adapted all existing ones by wrapping on registration which is
why this was easy.
* Left some notes in the code about where type resolution fits in.

## Related Issues

The issue is that all functions are dynamic .. they need not be so. We
also moved some dynamic behavior out of the planner and into the hot
path on recent refactors. This enables lowering during planning which is
a bit more powerful than even where we were before the refactor.

* Thoughts..
* I did this because we lost the ability to curry which was quite
powerful.
  * I want to rename ScalarFunctionFactory to ScalarFunction
  * I want to renamed ScalarUDF to ScalarFunctionImpl
  * These are internal names, so it doesn't really matter


## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)