feat(catalogs): Rust-based in-memory catalog and table (#4445)
## Changes Made

Move the implementation of MemoryCatalog and MemoryTable from Python
into Rust, thus making them the first full Rust-based implementations of
catalogs and tables!

In addition, added the ability to modify (append and overwrite)
MemoryTables.

Slightly related: changed Table ABC so that `append` and `overwrite`
need to be implemented instead of `write`.

## Related Issues

Follow-up to #4394

## Checklist

- [x] Documented in API Docs (if applicable)
- [x] Documented in User Guide (if applicable)
- [x] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [x] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)