refactor(exprs): temporals (#4446)
## Changes Made

Note: I did not move over the kernels because they are used for various
partitioning functions throughout core and other crates. To move the
kernels would be a bit bigger of a refactor that's out of scope for
getting everything converted to the new `ScalarUDF` methods.



## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)