feat(flotilla): Flotilla runner (#4459)
## Changes Made

Implement dedicated plan runner as a 0 cpu ray actor on the head node,
in order to support ray job + client mode.

Note: 
- This requires logical plan to be serializable, as such I added `SerDe`
to the Logical Plan. For scan operator, I added a custom de/serializer
that throws an error, because we should not be able to serialize scan
operators right now.
- There is no retries / fault tolerance on this plan runner actor. If it
dies, the query fails. This is the same behavior as the current ray
runner.

Additionally, this PR enables flotilla in CI. (I just replaced AQE with
flotilla).

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)