feat: add id and plan_id to otel subscribers, and add method for getting tree hierarchy for a pipeline (#4485)
## Changes Made

### - Creates a new method `get_relationship_info`

This currently isn't being used yet, but the general idea is that for
the OSS dashboard visualizations, we need some way to know information
about the plan structure (nodes and their children). Which we later fill
in with `RuntimeContext` information. `get_relationship_info` is purely
just to get the tree hierarchy.

### - Update the otel subscriber to include `id` as well as `plan_id`. 
As an added bonus, the otel subsciber now has access to some additional
fields that'll be useful for generic telemetry purposes (id, and
plan_id).

ÏÏ
## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)