chore: add additional context to the runtime stats events (#4520)
## Changes Made

adds an opaque `context` object that we can use to pass in additional
context when emitting the runtimestats events.

When using the new flotilla engine, this adds the following additional
properties: [plan_id, stage_id, node_id, task_id]


## Note for reviewers: (@colin-ho @srilman )

While this PR does introduce the extra information to associate an event
back to stages/plans/nodes, it does not yet give us the capability to
know the relationship between stages, or between nodes

For example, if we have a stage with 3 pipeline nodes, where (3) depends
on both (1) and (2), we have no way to know that as we're not tracking
the relationships. This is a very similar problem to
https://github.com/Eventual-Inc/Daft/pull/4485. In that case we were
able to just walk the tree and get the relationship up front, but with
our distributed plans we don't actually know what's going to happen
next, so we need to map the relationships as they are formed.


## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)