feat(flotilla): Actor pool project (#4497)
## Changes Made

Implement actor pool projects in flotilla.

### How it works
Actor pool projects are implemented as a separate distributed pipeline
node.
- They first spin up all necessary ray actors to run the UDFs, using
ray's scheduler to schedule them based on the resource requirements (num
cpus, num gpus, memory).
- Then, we group the ray actors by node id so we can figure out which
swordfish workers are colocated with which ray actors.
- The actor pool project pipeline node uses this information to schedule
tasks appropriately.
- If the input is a `task`, pick a new worker to run that task based on
round robin, append the actor pool project instruction (containing
references to the ray actor for this projection) and yield it.
- If the input is a `materialized output`, pick the worker to run this
task based on the location of the materialized output, append the actor
pool project instruction (containing references to the ray actor for
this projection) and yield it.
- Once all tasks are complete, spin down the actors.

### Notes
- Actor pool projects are pipelined with existing tasks if possible.
This was found to significantly reduce object store spilling. However,
this has implications for heterogenous clusters.
- Swordfish workers run one task at a time, and tasks are sized based on
the `max_sources_per_scan_task` config variable. This is not set in
stone and a more robust task sizing algorithm is necessary, and should
be discussed after this PR.

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)