feat: Url parse function (#4533)
## Changes Made

This PR implements a url parse expression. Note that I could not find a
way to implement the exact example given in #2951 with `col('url.*')`
but I used a struct field on which `.get("*")` can be called. Not sure
this is exactly what you had in mind @universalmind303?

## Related Issues

Closes #2951

## Checklist

- [x] Documented in API Docs (if applicable)
- [x] Documented in User Guide (if applicable)
- [x] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [x] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)