feat: Commit write sink (#4569)
## Changes Made

Adds a `CommitWrite` sink to swordfish, which performs overwrite /
overwrite partitions. In the future it will allow us to add other
functionality needed to do `commits`.

Mainly this PR is to allow flotilla to properly perform overwrites, as
previously it was doing it at a per-task level, which was incorrect.

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)