feat(flotilla): Flotilla progress bar (#4584)
## Changes Made

Since flotilla runs on a ray actor, we need to use the ray tqdm module.
The existing ray runner already uses it so its easy to let flotilla use
it as well.

It uses the same `make_or_update_total` model where the progress bar is
created if it does not exist, else the total is updated.

For this new flotilla progress bar, the totals are much more accurate
because the scheduler is able to get a full view of pending tasks,
whereas in the current ray runner it is limited to only tasks that are
dispatch-able.

However for tasks that depend on the output of previous tasks, such
downstream tasks from a limit, can still have fluctuating totals.



https://github.com/user-attachments/assets/e06dc3d8-4d75-4027-bcc5-efe168003193


## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)