chore: Improve codecov (#4601)
## Changes Made

Improve codecov, especially adding flotilla to be tracked by codecov.

This requires that we run the flotilla scheduler not on a ray actor,
because it the codecov doesn't reach the ray actor. So we use the
current model that we have, only run on ray actor if we are on ray
client mode. Otherwise it is ran locally, which should be already on the
ray head node.


## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)